+++ changelog.md
@@ -3,6 +3,7 @@
 ### v1.6.0
 * FEATURE | Added a message when player plays for the first time on the current version.
 * FIX | Prevented accidental opening of the links below the game screen.
+* FIX | Grade now updates on profile load.
 
 ### v1.5.2
 * CHANGE | Rebalanced grade requirements.